Move Button and Input components to global component.
